yt-download-python/requirements.txt
functions_framework
yt-dlp[default] @ https://github.com/yt-dlp/yt-dlp/archive/master.zip
google-cloud-storage
requests
